improved DownloadManager.replenish logging; refactoring
