+++ doc/architecture.txt
@@ -52,7 +52,8 @@ lowercase alphabet and numerals only; avoid capital letters and underscores.
 *	target.js	Request class that queries the current tab for a TARGET.
 *	albumlist.js	Request class that queries an ALBUMLIST for ALBUMS.
 *	album.js	Request class that queries an ALBUM for MEDIA.
-*	download.js	Internal download manager utility class.
+*	download.js	Download manager class.
+*	dlhelpers.js	Internal utilities for use by download.js.
 
 2.2	Panel code (data/content/panel/js/)
 -------------------------------------------
